### 2025-09-10 #724 @juanmarcosmoreno-arch
cowfootR. New submission
Nombre de la Persona Encargada: Juan Moreno
Usuario GitHub de la Persona Encargada: @juanmarcosmoreno-arch
Repositorio: https://github.com/juanmarcosmoreno-arch/cowfootR/
Versión Enviada: 0.1.0
Tipo de Envio: Estándar
Persona encargada de la edición: @maurolepore
Personas encargadas de la revisión: A definir, @paocorrales
Archivo: A definir
Versión Aceptada: A definir
Idioma: es
---

-   Pega el archivo DESCRIPTION completo dentro del siguiente bloque de código.

```
Package: cowfootR
Title: Tools to Estimate the Carbon Footprint of Dairy Farms
Version: 0.1.0
    person("Juan", "Moreno", email = "juanmarcosmoreno@gmail.com", role = c("aut", "cre"))
Description: Provides functions to calculate the carbon footprint of dairy farms based on based on methodologies of the International Dairy Federation and the Intergovernmental Panel on Climate Change methodologies. Includes tools for single-farm and batch analysis, report generation, and visualization.
Imports: writexl
    readxl,
    plotly,
    gt,
    tidyr

```

## Alcance

- Por favor, indica qué categoría(s) aplica(n) a este paquete. Las puedes encontrar en nuestras [políticas de inclusión de paquetes](https://devguide.ropensci.org/es/softwarereview_policies.es.html#package-categories). Por favor, tilda todas las apropiadas. Si tienes dudas, te sugerimos que comiences un pre-envío.

	- [ ] obtención de datos (data retrieval)
	- [ ] extracción de datos (data extraction)
	- [ ] manipulación de datos (data munging)
	- [ ] depósito de datos (data deposition)
    - [ ] validación y comprobación de datos (data validation and testing)
	- [ ] automatización de flujo de trabajo (workflow automation)
	- [ ] control de versiones (version control)
	- [ ] manejo de citas y bibliometría (citation management and bibliometrics)
	- [ ] capa de interfaz de software científico (scientific software wrappers)
	- [x] herramientas de reproducibilidad de campo y laboratorio (field and lab reproducibility tools)
	- [ ] enlaces a software de bases de datos (database software bindings)
	- [ ] datos geoespaciales (geospatial data)
	- [ ] tradución (translation)

- Explica cómo y por qué el paquete encaja dentro de estas categorías (1 a 3 oraciones):
cowfootR automatiza el flujo de trabajo para calcular huellas de carbono de tambos siguiendo IDF 2022 e IPCC 2019, con funciones modulares y un modo batch reproducible. Incluye chequeos básicos/validaciones de insumos y una viñeta que estandariza el procedimiento, favoreciendo la reproducibilidad en contextos de campo/laboratorio.
- ¿Cuál es la audiencia esperada y las aplicaciones científicas de este paquete?
Investigadores y profesionales en sustentabilidad agropecuaria, LCA/ACV, ganadería lechera, y actores de extensión/industria. Aplicaciones: evaluación comparativa (benchmarking) de tambos, análisis de mitigación, reportes de inventarios GHG, y apoyo a políticas/etiquetado ambiental.
- ¿Hay otros paquetes de R que logren el mismo objetivo? Si los hay, ¿cómo se diferencian del tuyo, o alcanzan nuestro [criterio del mejor de su categoría](https://devguide.ropensci.org/es/softwarereview_policies.es.html#overlap)?
No conocemos paquetes de R que implementen específicamente la norma IDF 2022 para el sector lechero (con FPCM e intensidades por ha) de forma integrada y reproducible. Existen herramientas LCA más generales o enfocadas a otros dominios, pero cowfootR se diferencia por su foco en tambo, el cumplimiento explícito de IDF/IPCC, y la automatización batch con métricas de intensidades (FPCM/ha) listas para informe.
-   (Si aplica) ¿Tu paquete cumple con nuestras [guías de _Ética, Privacidad de Datos e Investigación con Sujetos Humanos_](https://devguide.ropensci.org/es/softwarereview_policies.es.html#ethics-data-privacy-and-human-subjects-research)?
No aplica tratamiento de datos personales ni sensibles; el paquete opera con datos agregados de producción/insumos a nivel establecimiento. Cumple las guías de Ética y Privacidad indicadas (no recopila datos, no transmite a terceros).
- Si ya has hecho una consulta de pre-envío, por favor pega el enlace al _issue_ correspondiente, una publicación del foro, u otra discusión. Alternativamente, etiqueta a la persona del equipo editorial (con `@tag`) con la que te contactaste.
Aún no se realizó pre-envío
-   (Si aplica) Explica las razones por las cuales el paquete no satisface alguno de los [chequeos de `pkgcheck`](https://docs.ropensci.org/pkgcheck/).

## Revisiones Técnicas

Tilda los siguientes items para confirmar que los has completado:

- [X ] He leído la [guía de empaquetado de rOpenSci](https://devguide.ropensci.org/es/pkg_building.es.html).
- [X] He leído la [guía de autoría](https://devguide.ropensci.org/es/softwarereview_author.es.html) y pienso mantener este paquete durante al menos 2 años o encontrar un reemplazo.

Este paquete:

- [X] No viola los Términos de Servicio de ningún servicio con los que interactúa.
- [X] Tiene una licencia aceptada por CRAN y OSI.
- [ ] Contiene un [archivo README con instrucciones para instalar la versión de desarrollo](https://devguide.ropensci.org/es/pkg_building.es.html#readme).
- [X] Incluye [documentación con ejemplos para todas las funciones, creada con roxygen2](https://devguide.ropensci.org/es/pkg_building.es.html#documentation).
- [X] Contiene una viñeta (_vignette_) con ejemplos de sus funciones esenciales y su uso.
- [X] Tiene una [suite de tests](https://devguide.ropensci.org/es/pkg_building.es.html#testing).
- [X] Tiene [integración contínua](https://devguide.ropensci.org/es/pkg_ci.es.html), incluyendo reporte de cobertura de tests.

## Opciones de Publicación

- [X] ¿Tienes intenciones de publicar este paquete en CRAN?
- [ ] ¿Tienes intenciones de publicar este paquete en Bioconductor?

- [X] ¿Deseas enviar un Artículo de Aplicaciones sobre tu paquete a [Methods in Ecology and Evolution (documento en Inglés)](http://besjournals.onlinelibrary.wiley.com/hub/journal/10.1111/(ISSN)2041-210X/)? Si es así:

<summary>Opciones para MEE</summary>

- [X] Este paquete es novedoso y será de interés para la mayoría de las persona que leen la revista.
- [X] El manuscrito que describe el paquete no tiene más de 3000 palabras y está escrito en Inglés.
- [X] Tienes intenciones de archivar el código del paquete en un repositorio a largo plazo, que cumple los requerimientos de la revista (mira las [Políticas de Publicación de MEE (documento en Inglés)](http://besjournals.onlinelibrary.wiley.com/hub/journal/10.1111/(ISSN)2041-210X/journal-resources/policy-on-publishing-code.html))
- (*Alcance: Considera los [Objetivos y Alcance de MEE (documento en Inglés)](https://besjournals.onlinelibrary.wiley.com/hub/journal/2041210x/policyonpublishingcode.html) para tu manuscrito. No otorgamos garatías de que tu manuscrito esté en el ámbito de MEE.*)
- (*Aunque no es requerido, recomendamos tener un manuscrito completamente preparado y en Inglés, al momento de enviar el paquete.*)
- (*Por favor, no envíes tu paquete de forma separada a Methods in Ecology and Evolution*)

## Código de Conducta

- [X] Estoy de acuerdo en cumplir el [Código de Conducta de rOpenSci](https://devguide.ropensci.org//es/softwarereview_policies.es.html#code-of-conduct) durante el proceso de revisión, y en mantener mi paquete si éste es aceptado.
