### 2025-09-10 63599a9
Deploying to gh-pages from @ juanmarcosmoreno-arch/cowfootR@636f140399de200dd8b35c8ca7b9e4fc95025327 🚀
