### 2025-09-10 f84e76b
CI: build vignettes in check & ensure vignette deps
