+++ .github/workflows/R-CMD-check.yaml
@@ -48,5 +48,5 @@ jobs:
       - uses: r-lib/actions/check-r-package@v2
         with:
           upload-snapshots: true
-          build-vignettes: false
+          build-vignettes: true
           args: 'c("--no-manual")'
