+++ tests/testthat/test-intensities.R
@@ -1,0 +1,148 @@
+testthat::test_that("setup: intensities functions exist", {
+  testthat::skip_if_not(exists("calc_intensity_area"))
+  testthat::skip_if_not(exists("calc_intensity_litre"))
+})
+
+# -------------------------------------------------------------------------
+# calc_intensity_area tests
+# -------------------------------------------------------------------------
+
+testthat::test_that("calc_intensity_area: basic case returns finite non-negative intensity", {
+  fn <- calc_intensity_area
+
+  # Emissions total and area values for test
+  total <- 1000
+  area  <- 50
+
+  # Try different calling styles depending on function signature
+  out <- try(safe_call(fn,
+                       canonical_args = list(total_CO2eq = total, area_ha = area),
+                       positional_args = list(total, area),
+                       df_args = list(total_CO2eq = total, area_ha = area)
+  ), silent = TRUE)
+
+  if (inherits(out, "try-error")) testthat::skip("Function signature incompatible — skipping test.")
+
+  val <- pick_named_numeric(out, patterns = c("per_ha", "intens", "ha"))
+  if (length(val) == 0) testthat::skip("No intensity field detected in output.")
+
+  testthat::expect_true(all(is.finite(val)))
+  testthat::expect_true(all(val >= 0))
+})
+
+testthat::test_that("calc_intensity_area: higher emissions with same area → higher intensity", {
+  fn <- calc_intensity_area
+
+  o1 <- try(safe_call(fn,
+                      canonical_args = list(total_CO2eq = 1000, area_ha = 100),
+                      positional_args = list(1000, 100),
+                      df_args = list(total_CO2eq = 1000, area_ha = 100)
+  ), silent = TRUE)
+
+  o2 <- try(safe_call(fn,
+                      canonical_args = list(total_CO2eq = 1100, area_ha = 100),
+                      positional_args = list(1100, 100),
+                      df_args = list(total_CO2eq = 1100, area_ha = 100)
+  ), silent = TRUE)
+
+  if (inherits(o1, "try-error") || inherits(o2, "try-error")) testthat::skip("Signature incompatible — skipping monotonicity test.")
+
+  v1 <- pick_named_numeric(o1, patterns = c("per_ha", "intens", "ha"))
+  v2 <- pick_named_numeric(o2, patterns = c("per_ha", "intens", "ha"))
+  if (length(v1) == 0 || length(v2) == 0) testthat::skip("No intensity field detected.")
+
+  testthat::expect_true(is.finite(mean(v1)) && is.finite(mean(v2)))
+  testthat::expect_gt(mean(v2), mean(v1))
+})
+
+testthat::test_that("calc_intensity_area: zero or NA area handled gracefully", {
+  fn <- calc_intensity_area
+
+  # Zero area: should return NA/Inf or error
+  o0 <- try(safe_call(fn,
+                      canonical_args = list(total_CO2eq = 100, area_ha = 0),
+                      positional_args = list(100, 0),
+                      df_args = list(total_CO2eq = 100, area_ha = 0)
+  ), silent = TRUE)
+
+  if (!inherits(o0, "try-error")) {
+    v0 <- pick_named_numeric(o0, patterns = c("per_ha", "intens", "ha"))
+    testthat::expect_true(any(is.na(v0) | is.infinite(v0)))
+  } else testthat::expect_true(TRUE)
+
+  # NA area: should return NA or error
+  oNA <- try(safe_call(fn,
+                       canonical_args = list(total_CO2eq = 100, area_ha = NA_real_),
+                       positional_args = list(100, NA_real_),
[…]

[…]
