### 2025-09-11 b4d831f
Update package after CRAN-ready check
