+++ tests/testthat.R
@@ -3,3 +3,7 @@ library(testthat)
 library(cowfootR)
 
 test_check("cowfootR")
+
+if (!requireNamespace("withr", quietly = TRUE)) {
+  skip("withr not available for tests")
+}
