+++ vignettes/IPCC_Methodology_Tiers_cowfootR.Rmd
@@ -10,6 +10,7 @@ vignette: >
 
 ```{r setup, include = FALSE}
 knitr::opts_chunk$set(
+  fig.alt = "Figura generada por la viñeta; ver texto para detalles.",
   collapse = TRUE,
   comment = "#>",
   fig.width = 8,
