+++ vignettes/Single_Farm_Analysis_cowfootR.Rmd
@@ -11,6 +11,7 @@ vignette: >
 `
 ```{r setup, include = FALSE}
 knitr::opts_chunk$set(
+  fig.alt = "Figura generada por la viñeta; ver texto para detalles.",
   collapse = TRUE,
   comment = "#>",
   fig.width = 7,
