+++ vignettes/Introduction_to_Dairy_LCA_cowfootR.Rmd
@@ -9,6 +9,7 @@ vignette: >
 
 ```{r setup, include = FALSE}
 knitr::opts_chunk$set(
+  fig.alt = "Figura generada por la viñeta; ver texto para detalles.",
   collapse = TRUE,
   comment = "#>",
   fig.width = 7,
