### 2025-10-07 4e15c1a
Update paper.md
