### 2025-10-07 2049101
Deploying to gh-pages from @ juanmarcosmoreno-arch/cowfootR@4e15c1ac1003912a8033e866ee87ca0eae87e63e 🚀
