### 2025-10-16 2710dab
Deploying to gh-pages from @ juanmarcosmoreno-arch/cowfootR@2d95bc4b570bda77b4689eeefec5eea9b9a0dc51 🚀
