+++ pkgdown.yml
@@ -7,7 +7,7 @@ articles:
   IPCC_Methodology_Tiers_cowfootR: IPCC_Methodology_Tiers_cowfootR.html
   Single_Farm_Analysis_cowfootR: Single_Farm_Analysis_cowfootR.html
   workflow: workflow.html
-last_built: 2025-10-07T21:10Z
+last_built: 2025-10-16T18:34Z
 urls:
   reference: https://juanmarcosmoreno-arch.github.io/cowfootR/reference
   article: https://juanmarcosmoreno-arch.github.io/cowfootR/articles
