### 2025-10-16 76835ba
Deploying to gh-pages from @ juanmarcosmoreno-arch/cowfootR@d5fe570ad112e1c2d77e617771925f73ef5d0ceb 🚀
