### 2025-10-18 #724 @maurolepore
@ldecicco-USGS, following our brief exchange I'm happy to handle this submission but before I self assign I'd like to discuss out categories.

@juanmarcosmoreno-arch, estuve revisando las categorías que propusiste y me gustaría conversar un poco con vos para entender mejor cómo encaja cowfootR en el scope de rOpenSci.

Me quedé pensando en las dos categorías que seleccionaste:

**Obtención de datos:** Por lo que veo en el código, me parece que el paquete calcula huellas de carbono a partir de datos que el usuario provee (archivos Excel, data frames), más que obtener datos de fuentes online. ¿Estás de acuerdo con eso? Quizás me estoy perdiendo algo de cómo funciona el paquete?

**Automatización de flujo de trabajo:** Veo que `calc_batch()` permite procesamiento por lotes, lo cual es super útil pero no estoy del todo seguro si esto encaja con lo que rOpenSci entiende por "automatización de flujo de trabajo" (como drake o targets que orquestan diferentes tareas computacionales). ¿Vos qué pensás?

**Idea alternativa - Herramientas de reproducibilidad de campo y laboratorio:** Creo que esta categoría podría encajar mejor. Me parece que cowfootR hace varias cosas que se alinean con esto. Vos mismo mencionaste "favoreciendo la reproducibilidad en contextos de campo/laboratorio" en tu envío, lo cual me hizo preguntarme si era una categoría que habías considerado?

Quisiera también discutirlo con el equipo editorial, pero antes me gustaría saber tu opinión.

---

**Definiciones de rOpenSci (de las [políticas de revisión](https://devguide.ropensci.org/es/softwarereview_policies.es.html#package-categories)):**

> **obtención de datos**: Paquetes para acceder y descargar datos de fuentes en línea con usos científicos. Nuestra definición de "usos científicos" es amplia. Incluye servicios de almacenamiento de datos, revistas y otros servidores remotos, ya que existe una gran variedad de fuentes de datos que pueden ser de interés para quienes hacen ciencia. Sin embargo, los paquetes de obtención de datos deben centrarse en las *fuentes* o *temas* de los datos y no en *servicios*.

> **automatización del flujo de trabajo**: Herramientas que automatizan y encadenan flujos de trabajo, como los sistemas de construcción y las herramientas para gestionar la integración continua. No incluye las herramientas generales para la programación literiaria (por ejemplo, las extensiones de R markdown que no están en los temas anteriores). (Ejemplo: [**drake**](https://github.com/ropensci/software-review/issues/156))

> **herramientas de reproducibilidad de campo y laboratorio**: Paquetes que mejoran la reproducibilidad de los flujos de trabajo del mundo real mediante la estandarización y la automatización de los protocolos de campo y de laboratorio. Por ejemplo, el seguimiento y el etiquetado de las muestras, la generación de formularios y hojas de datos, la interconexión con los equipos de laboratorio o los sistemas de información, y la ejecución de diseños experimentales. (Ejemplo: [**baRcodeR**](https://github.com/ropensci/software-review/issues/336))
