### 2025-11-02 #724 @maurolepore
Gracias @juanmarcosmoreno-arch,

Acabo de modificar el primer comentario para reflejar el cambio.

Es un gusto acompañarte a lo largo del proceso en el rol de editor. Para ayudarte a rastrear mis comentarios, los etiqueto con "ml" y una secuencia numérica: ml01, ml02, etc:

* Los comentarios iniciados con "bullet" (como este) sugieren cosas que no requieren tu acción o respuesta.
- [ ] Los comentarios iniciados con "checkbox"  (como este) piden cosas que requieren tu acción y respuesta.

Acá van mis primeros comentarios:

- [ ] ml01. ¿Confirmas que la revisión es 100% en Castellano? (Con [karel](https://github.com/ropensci/software-review/issues/620) hicimos una revisión mixta entre Castellano e Ingles.)

- [ ] ml02. Por favor sugerí 3 revisores/as. Aunque usaremos no mas que uno, tu sugerencia sirve para ver que tipo de experiencia sería útil[^1].

Estoy por comenzar mis chequeos iniciales[^2]. Pronto los voy a publicar aca.

> [^1]: Como responsable de la edición, utiliza las posibles sugerencias realizadas por quienes presentaron el paquete (aunque éstas pueden tener una visión limitada de los tipos de conocimientos necesarios; sugerimos no utilizar más de una de las personas sugeridas);
> -- https://devdevguide.netlify.app/es/softwarereview_editor.es#where-to-look-for-reviewers

> [^2]: Después de publicar las comprobaciones automáticas, utiliza la plantilla editorial para guiar las comprobaciones iniciales y registrar tu respuesta sobre el envío. También puedes agilizar las comprobaciones de edición utilizando el paquete pkgreviewr creado por la editora asociada Anna Krystalli. Procura terminar las comprobaciones y empezar a buscar personas para hacer la revisión en un plazo de 5 días laborables.
> -- https://devdevguide.netlify.app/es/softwarereview_editor.es#upon-submission
