### 2025-11-10 cd97d05
Update paper.md
