+++ paper.md
@@ -46,25 +46,24 @@ attributed to a particular product or process, expressed in terms of the carbon
 dioxide equivalent (CO₂e or CO₂eq). As far as milk production is concerned, the 
 carbon footprint includes emissions from, e.g., enteric fermentation, fertiliser 
 management, feed production, use of outside inputs, and energy consumption 
-(Stolarski et al., 2025).
+[@stolarski2025].
 The dairy industry contributes approximately 4% of global greenhouse gas 
 emissions, with carbon footprint values ranging from 0.78 to 3.20 kg CO₂eq kg⁻¹ 
-of milk across different production systems (Flysjö et al., 2011; 
-Stolarski et al., 2025). The Intergovernmental Panel on Climate Change 
-emphasizes that livestock production systems require accurate quantification 
-methods to support effective mitigation strategies and policy development 
-(IPCC, 2019). Similarly, the International Dairy Federation has established 
-comprehensive guidelines for standardized carbon footprint assessment, 
+of milk across different production systems [@flysjo2011; @stolarski2025]. 
+The Intergovernmental Panel on Climate Change emphasizes that livestock production 
+systems require accurate quantification methods to support effective mitigation 
+strategies and policy development [@ipcc2019]. Similarly, the International Dairy 
+Federation has established comprehensive guidelines for standardized carbon footprint assessment, 
 recognizing the critical need for consistent methodologies that enable fair 
 comparison across different dairy systems while accounting for regional 
-variations (IDF, 2022).
+variations [@idf2022].
 With increasing regulatory pressure from initiatives like the EU Green Deal and 
 Corporate Sustainability Reporting Directive, there is urgent need for 
 standardized, accessible tools to quantify dairy farm carbon footprints 
-(The European Parliament and of the Council, 2022). Current life cycle 
+[@eu2022]. Current life cycle 
 assessment (LCA) software solutions have significant limitations: most are 
 expensive commercial packages requiring specialized training, methodological 
-inconsistencies limit result comparability (Pirlo, 2012), and many lack 
+inconsistencies limit result comparability [@pirlo2012], and many lack 
 transparency or regional adaptation capabilities. These barriers prevent 
 widespread adoption of standardized practices, particularly among smaller farms 
 and developing regions.
@@ -87,7 +86,8 @@ and computing intensity metrics.
 
 # Workflow
 
-![Fig.1 Workflow\label{fig:wf}](man/figures/cowfoot_workflow.png)
+![Workflow of the cowfootR package.](man/figures/cowfoot_workflow.png)
+
 
 # Availability
 
