### 2025-11-10 43d14f2
Deploying to gh-pages from @ juanmarcosmoreno-arch/cowfootR@cd97d05a93e070a0269c386937d2ab360cbdb781 🚀
