+++ reference/calc_emissions_enteric.md
@@ -1,0 +1,247 @@
+# Calculate enteric methane emissions
+
+Estimates enteric methane (CH4) emissions from cattle using IPCC Tier 1
+or Tier 2 approaches with practical defaults for dairy systems.
+
+## Usage
+
+``` r
+calc_emissions_enteric(
+  n_animals,
+  cattle_category = "dairy_cows",
+  production_system = "mixed",
+  avg_milk_yield = 6000,
+  avg_body_weight = NULL,
+  dry_matter_intake = NULL,
+  feed_inputs = NULL,
+  ym_percent = 6.5,
+  emission_factor_ch4 = NULL,
+  tier = 1L,
+  gwp_ch4 = 27.2,
+  boundaries = NULL
+)
+```
+
+## Arguments
+
+- n_animals:
+
+  Numeric scalar \> 0. Number of animals.
+
+- cattle_category:
+
+  Character. One of "dairy_cows", "heifers", "calves", "bulls". Default
+  = "dairy_cows".
+
+- production_system:
+
+  Character. One of "intensive", "extensive", "mixed". Default =
+  "mixed".
+
+- avg_milk_yield:
+
+  Numeric \>= 0. Average annual milk yield per cow (kg/year). Default
+  = 6000. Used in Tier 2 fallback for dairy cows.
+
+- avg_body_weight:
+
+  Numeric \> 0. Average live weight (kg). If NULL, a category-specific
+  default is used (e.g. 550 kg for dairy cows).
+
+- dry_matter_intake:
+
+  Numeric \> 0. Dry matter intake (kg/animal/day). If provided (Tier 2),
+  overrides body-weight/energy-based estimation.
+
+- feed_inputs:
+
+  Named numeric vector/list with feed DM amounts in kg/year per herd
+  (e.g., grain_dry, grain_wet, byproducts, proteins). Optional. If given
+  and `dry_matter_intake` is NULL, DMI is inferred as
+  `sum(feed_inputs)/(n_animals*365)`.
+
+- ym_percent:
+
+  Numeric in (0, 100\]. Methane conversion factor Ym (% of GE to CH4).
+  Default = 6.5.
+
+- emission_factor_ch4:
+
+  Numeric \> 0. If provided, CH4 EF (kg CH4/head/year) is used directly;
+  otherwise it is calculated (Tier 1 or Tier 2).
+
+- tier:
+
+  Integer 1 or 2. Default = 1.
+
+- gwp_ch4:
+
+  Numeric. GWP for CH4 (100-yr, AR6). Default = 27.2.
+
+- boundaries:
+
+  Optional list from
+  [`set_system_boundaries()`](https://juanmarcosmoreno-arch.github.io/cowfootR/reference/set_system_boundaries.md).
+
+## Value
+
+List with CH4 (kg), CO2eq (kg), inputs, factors, and metadata. Includes
+`co2eq_kg` for compatibility with
+[`calc_total_emissions()`](https://juanmarcosmoreno-arch.github.io/cowfootR/reference/calc_total_emissions.md).
+
+## Examples
+
+``` r
+# \donttest{
+# Tier 1, mixed dairy cows
+calc_emissions_enteric(n_animals = 100)
+#> $source
+#> [1] "enteric"
+#> 
+#> $category
[…]

[…]
