### 2025-11-10 2fc44e2
Update paper.md
