### 2025-11-10 5c3ccec
Deploying to gh-pages from @ juanmarcosmoreno-arch/cowfootR@092ac751c9a3b6a493a60b64899c4984d1bb16ad 🚀
