### 2025-11-10 f9d8ac6
Deploying to gh-pages from @ juanmarcosmoreno-arch/cowfootR@2fc44e2cbd24548cbfa39c0ea6ccac5fc15f2d1f 🚀
