### 2025-11-10 353d16c
Update paper.md
