+++ paper.md
@@ -91,8 +91,10 @@ and computing intensity metrics.
 
 # Availability
 
-cowfootR package is available on GitHub 
-(https://github.com/juanmarcosmoreno/cowfootR). Documentation, including 
+The cowfootR package is freely available on both [CRAN](https://cran.r-project.org/package=cowfootR) and 
+[GitHub](https://github.com/juanmarcosmoreno-arch/cowfootR). Comprehensive documentation, including vignettes and 
+reproducible examples, is provided to facilitate adoption and integration into research and sustainability assessment workflows.
+cowfootR package is available on GitHub (https://github.com/juanmarcosmoreno-arch/cowfootR). Documentation, including 
 vignettes and examples, is provided to facilitate adoption.
 
 # Acknowledgements
