### 2025-11-10 670005d
Deploying to gh-pages from @ juanmarcosmoreno-arch/cowfootR@353d16c43ac47db9a99bba548838320a832a131a 🚀
