### 2025-11-15 #724 @maurolepore
<details><summary>EN</summary>

For the record, I'm attempting some housekeeping tasks.

In our [reviews dashboard](https://dashboard.ropensci.org/reviews.html) I see "The following issue has multiple stages: 724" so I removed the label `3/reviewer(s)-assigned` to see if that fixes the warning. I'm not sure since only 1 but not 2 reviewers have been assegined so I guess this intermediate state could be interpreted as both `2/seeking-reviewer(s)` and also `3/reviewer(s)-assigned`.

I don't recall having this dilemma in other issues.

Para que conste, estoy intentando hacer algunas tareas de mantenimiento.

En nuestro [tablero de revisiones](https://dashboard.ropensci.org/reviews.html) veo “The following issue has multiple stages: 724”, así que quité la etiqueta "3/reviewer(s)-assigned" para ver si eso soluciona la advertencia. No estoy seguro, ya que solo 1 y no 2 revisore/as han sido asignados, así que supongo que este estado intermedio podría interpretarse tanto como "2/seeking-reviewer(s)" como "3/reviewer(s)-assigned".

No recuerdo haber tenido este dilema en otros issues.
