### 2025-12-10 #724 @NPJuncal
Hola @maurolepore . He empezado con mi revision y tengo una duda. Hay un manuscrito asociado con este paquete que tengo que revisar o tan solo reviso el codigo y el repositorio (https://juanmarcosmoreno-arch.github.io/cowfootR/)?
