### 2025-12-29 #724 @maurolepore
@juanmarcosmoreno-arch, solo para reconocer que debido a las fiestas esperaria demora en tu respuesta. En cuanto vuelvas a trabajar por favor ponenos al tanto de tu plan.

Felicidades!
