### 2025-12-30 #724 @juanmarcosmoreno-arch
@maurolepore, una consulta, como se deben responder los comentarios de los revisores? solo hagos lso cambios y subo a github, o se debe enviar una respuesta de alguna otra forma comentario por comentario. Gracias
