### 2025-12-30 573ab74
Create CODE_OF_CONDUCT.md
