+++ .github/CODE_OF_CONDUCT.md
@@ -1,0 +1,14 @@
+# Code of Conduct
+
+This project follows the rOpenSci Code of Conduct.
+
+All contributors, maintainers, and users of this project are expected to
+adhere to the rOpenSci Code of Conduct to ensure a welcoming and inclusive
+environment.
+
+For more information, please see:
+https://ropensci.org/code-of-conduct/
+
+Instances of abusive, harassing, or otherwise unacceptable behavior may be
+reported to the rOpenSci team via the reporting guidelines described at the
+link above.
