### 2025-12-30 34b5115
Deploying to gh-pages from @ juanmarcosmoreno-arch/cowfootR@adc890253ea2a1c86e92e41be3c9f03ccda7bdfc 🚀
