+++ reference/calc_emissions_inputs.md
@@ -348,7 +348,7 @@ calc_emissions_inputs(conc_kg = 1000, fert_n_kg = 200, region = "EU")
 #> [1] "IDF 2022; generic LCI sources"
 #> 
 #> $date
-#> [1] "2025-11-10"
+#> [1] "2025-12-30"
 #> 
 
 # With uncertainty analysis
@@ -598,6 +598,6 @@ calc_emissions_inputs(feed_corn_kg = 2000, region = "US", include_uncertainty = 
 #> [1] "IDF 2022; generic LCI sources"
 #> 
 #> $date
-#> [1] "2025-11-10"
+#> [1] "2025-12-30"
 #> 
 ```
