+++ reference/calc_emissions_soil.md
@@ -168,7 +168,7 @@ calc_emissions_soil(
 #> [1] "IPCC 2019 Refinement, IDF 2022"
 #> 
 #> $date
-#> [1] "2025-11-10"
+#> [1] "2025-12-30"
 #> 
 #> $per_hectare_metrics
 #> $per_hectare_metrics$n_input_kg_per_ha
@@ -279,7 +279,7 @@ calc_emissions_soil(n_fertilizer_synthetic = 2000, include_indirect = FALSE)
 #> [1] "IPCC 2019 Refinement, IDF 2022"
 #> 
 #> $date
-#> [1] "2025-11-10"
+#> [1] "2025-12-30"
 #> 
 #> $source_contributions
 #> $source_contributions$synthetic_fertilizer_pct
