+++ articles/Introduction_to_Dairy_LCA_cowfootR.md
@@ -236,7 +236,7 @@ print(enteric_emissions)
 #> [1] "IPCC 2019 Refinement, IDF 2022"
 #> 
 #> $date
-#> [1] "2025-11-10"
+#> [1] "2025-12-30"
 #> 
 #> $per_animal
 #> $per_animal$ch4_kg
@@ -330,7 +330,7 @@ print(manure_emissions)
 #> [1] "IPCC 2019 Refinement, IDF 2022"
 #> 
 #> $date
-#> [1] "2025-11-10"
+#> [1] "2025-12-30"
 #> 
 #> $per_cow
 #> $per_cow$ch4_kg
@@ -441,7 +441,7 @@ print(soil_emissions)
 #> [1] "IPCC 2019 Refinement, IDF 2022"
 #> 
 #> $date
-#> [1] "2025-11-10"
+#> [1] "2025-12-30"
 #> 
 #> $per_hectare_metrics
 #> $per_hectare_metrics$n_input_kg_per_ha
@@ -565,7 +565,7 @@ print(energy_emissions)
 #> [1] "IPCC 2019 Refinement, IDF 2022"
 #> 
 #> $date
-#> [1] "2025-11-10"
+#> [1] "2025-12-30"
 #> 
 #> $energy_metrics
 #> $energy_metrics$electricity_share_pct
@@ -796,7 +796,7 @@ print(input_emissions)
 #> [1] "IDF 2022; generic LCI sources"
 #> 
 #> $date
-#> [1] "2025-11-10"
+#> [1] "2025-12-30"
 ```
 
 #### Step 4: Aggregate Total Emissions
@@ -824,7 +824,7 @@ total_emissions
 #>   manure : 159755.4 kg CO2eq
 #>   soil : 69658.88 kg CO2eq
 #> 
-#> Calculated on: 2025-11-10
+#> Calculated on: 2025-12-30
 ```
 
 #### Step 5: Calculate Intensity Metrics
@@ -853,7 +853,7 @@ print(milk_intensity)
 #>  Protein content: 3.2 %
 #> 
 #> Total emissions: 645,948 kg CO2eq
-#> Calculated on: 2025-11-10
+#> Calculated on: 2025-12-30
 ```
 
 ##### Area Intensity
@@ -890,7 +890,7 @@ print(area_intensity)
 #>  infrastructure: 5.0 ha (4.2%) -> 26914 kg CO2eq
 #> 
 #> Total emissions: 645,948 kg CO2eq
-#> Calculated on: 2025-11-10
+#> Calculated on: 2025-12-30
 ```
 
 ### Visualizing Results
