### 2025-12-30 3b946db
Deploying to gh-pages from @ juanmarcosmoreno-arch/cowfootR@573ab74ce282c59b3432df285e9ee9a27d280492 🚀
