### 2025-12-30 caeb260
Deploying to gh-pages from @ juanmarcosmoreno-arch/cowfootR@b7bbeb6d5a140dadf6b411448514877ae9899d23 🚀
