### 2025-12-30 78501d9
Update Introduction_to_Dairy_LCA_cowfootR.Rmd
