+++ vignettes/Introduction_to_Dairy_LCA_cowfootR.Rmd
@@ -1,8 +1,8 @@
 ---
-title: "Introduction_to_Dairy_LCA_cowfootR"
+title: "Introduction to Dairy Life Cycle Assessment with cowfootR"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{Introduction_to_Dairy_LCA_cowfootR}
+  %\VignetteIndexEntry{01 - Introduction to Dairy Life Cycle Assessment}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
@@ -31,6 +31,18 @@ library(dplyr)
 The dairy industry plays a crucial role in global food security, but it also contributes significantly to greenhouse gas emissions. Understanding and quantifying the environmental impact of dairy production is essential for sustainable development and climate change mitigation.
 
 The **cowfootR** package provides a comprehensive toolkit for calculating dairy farm carbon footprints following internationally recognized standards, specifically the International Dairy Federation (IDF) 2022 guidelines and IPCC 2019 methodologies.
+
+## How to Use This Vignette
+
+This vignette is designed as a step-by-step tutorial for new users of
+**cowfootR**. It introduces the concepts of dairy life cycle assessment (LCA)
+and demonstrates how to calculate greenhouse gas emissions for a single farm.
+
+You can:
+- Read it sequentially as a guided example, or
+- Jump directly to the sections of interest (e.g., emissions, intensities, or visualization).
+
+All examples use simplified, hypothetical data intended for learning purposes.
 
 ## Theoretical Background
 
@@ -81,6 +93,20 @@ install.packages("cowfootR")
 ```{r}
 library(cowfootR)
 ```
+### Input Data Structure
+
+Most `cowfootR` functions expect farm information either as:
+- Individual numeric arguments (e.g. number of animals, litres of milk), or
+- A structured list containing farm characteristics.
+
+In this vignette, we use a simple list (`farm_data`) to keep all farm-related
+information together. This approach improves readability and makes it easier
+to reuse the same data across multiple calculation steps.
+
+Most emission functions return a **list** containing:
+- Total emissions for that source (kg CO₂eq)
+- A breakdown by gas or process
+- Metadata describing the calculation method
 
 ### Basic Workflow
 
@@ -91,6 +117,34 @@ The typical cowfootR workflow involves four main steps:
 3. **Aggregate total emissions**
 4. **Calculate intensity metrics**
 
+
+## Data Requirements: Required vs Optional
+
+### Required Data
+
+The following information is required to run the core cowfootR functions and
+perform a basic farm-level carbon footprint assessment:
+
+- Herd size and composition
+- Annual milk production
+- Total farm area
+- Major nitrogen inputs (fertilizer or excreta)
+
+### Optional but Recommended Data
+
+Providing additional farm-specific information improves the accuracy and
+interpretability of results:
+
+- Detailed feed quantities and composition
+- Animal productivity parameters (e.g. milk yield per cow)
+- Energy use disaggregated by source
+- Soil type and climate characteristics
+- Allocation of land use by category
+
+If some optional data are not available, cowfootR applies default values based
+on IPCC and IDF guidance. However, users are encouraged to provide farm-specific
+data whenever possible.
+
 Let's walk through a simple example:
 
 ## Example: Basic Farm Assessment
@@ -100,7 +154,7 @@ Let's walk through a simple example:
 ```{r}
 # Define farm-gate boundaries (most common approach)
 boundaries <- set_system_boundaries("farm_gate")
-print(boundaries)
+boundaries
 ```
 
 ### Step 2: Basic Farm Data
@@ -130,7 +184,7 @@ farm_data <- list(
[…]
