### 2025-12-30 1b7eee6
Deploying to gh-pages from @ juanmarcosmoreno-arch/cowfootR@78501d9e278e9cbf581edb941bae9b741a5da2fd 🚀
