+++ articles/index.md
@@ -3,7 +3,8 @@
 ### All vignettes
 
 - [Complete_Parameter_Reference_Guide](https://juanmarcosmoreno-arch.github.io/cowfootR/articles/Complete_Parameter_Reference_Guide.md):
-- [Introduction_to_Dairy_LCA_cowfootR](https://juanmarcosmoreno-arch.github.io/cowfootR/articles/Introduction_to_Dairy_LCA_cowfootR.md):
+- [Introduction to Dairy Life Cycle Assessment with
+  cowfootR](https://juanmarcosmoreno-arch.github.io/cowfootR/articles/Introduction_to_Dairy_LCA_cowfootR.md):
 - [IPCC_Methodology_Tiers_cowfootR](https://juanmarcosmoreno-arch.github.io/cowfootR/articles/IPCC_Methodology_Tiers_cowfootR.md):
 - [Single_Farm_Analysis_cowfootR](https://juanmarcosmoreno-arch.github.io/cowfootR/articles/Single_Farm_Analysis_cowfootR.md):
 - [cowfootR
