+++ articles/Introduction_to_Dairy_LCA_cowfootR.md
@@ -1,4 +1,4 @@
-# Introduction_to_Dairy_LCA_cowfootR
+# Introduction to Dairy Life Cycle Assessment with cowfootR
 
 ## Introduction to Dairy Life Cycle Assessment
 
@@ -14,6 +14,20 @@ The **cowfootR** package provides a comprehensive toolkit for
 calculating dairy farm carbon footprints following internationally
 recognized standards, specifically the International Dairy Federation
 (IDF) 2022 guidelines and IPCC 2019 methodologies.
+
+### How to Use This Vignette
+
+This vignette is designed as a step-by-step tutorial for new users of
+**cowfootR**. It introduces the concepts of dairy life cycle assessment
+(LCA) and demonstrates how to calculate greenhouse gas emissions for a
+single farm.
+
+You can: - Read it sequentially as a guided example, or - Jump directly
+to the sections of interest (e.g., emissions, intensities, or
+visualization).
+
+All examples use simplified, hypothetical data intended for learning
+purposes.
 
 ### Theoretical Background
 
@@ -74,6 +88,21 @@ install.packages("cowfootR")
 library(cowfootR)
 ```
 
+#### Input Data Structure
+
+Most `cowfootR` functions expect farm information either as: -
+Individual numeric arguments (e.g. number of animals, litres of milk),
+or - A structured list containing farm characteristics.
+
+In this vignette, we use a simple list (`farm_data`) to keep all
+farm-related information together. This approach improves readability
+and makes it easier to reuse the same data across multiple calculation
+steps.
+
+Most emission functions return a **list** containing: - Total emissions
+for that source (kg CO₂eq) - A breakdown by gas or process - Metadata
+describing the calculation method
+
 #### Basic Workflow
 
 The typical cowfootR workflow involves four main steps:
@@ -82,6 +111,33 @@ The typical cowfootR workflow involves four main steps:
 2.  **Calculate emissions by source**
 3.  **Aggregate total emissions**
 4.  **Calculate intensity metrics**
+
+### Data Requirements: Required vs Optional
+
+#### Required Data
+
+The following information is required to run the core cowfootR functions
+and perform a basic farm-level carbon footprint assessment:
+
+- Herd size and composition
+- Annual milk production
+- Total farm area
+- Major nitrogen inputs (fertilizer or excreta)
+
+#### Optional but Recommended Data
+
+Providing additional farm-specific information improves the accuracy and
+interpretability of results:
+
+- Detailed feed quantities and composition
+- Animal productivity parameters (e.g. milk yield per cow)
+- Energy use disaggregated by source
+- Soil type and climate characteristics
+- Allocation of land use by category
+
+If some optional data are not available, cowfootR applies default values
+based on IPCC and IDF guidance. However, users are encouraged to provide
+farm-specific data whenever possible.
 
 Let’s walk through a simple example:
 
@@ -92,7 +148,7 @@ Let’s walk through a simple example:
 ``` r
 # Define farm-gate boundaries (most common approach)
 boundaries <- set_system_boundaries("farm_gate")
-print(boundaries)
+boundaries
 #> $scope
 #> [1] "farm_gate"
 #> 
@@ -127,7 +183,7 @@ farm_data <- list(
   electricity_kwh = 35000   # Annual electricity use
 )
 
-print(farm_data)
+farm_data
 #> $dairy_cows
 #> [1] 100
 #> 
@@ -169,6 +225,15 @@ calculation functions:
 
 ##### Enteric Fermentation
 
+Enteric fermentation is typically the largest source of emissions in
+dairy systems. The function
[…]
