### 2025-12-30 0e6fe38
Deploying to gh-pages from @ juanmarcosmoreno-arch/cowfootR@98bdc2204aff6154e98abd6a669ba2fb1df3d1e0 🚀
