+++ articles/Single_Farm_Analysis_cowfootR.md
@@ -1,18 +1,16 @@
-# Single_Farm_Analysis_cowfootR
-
-\`
+# Single Farm Analysis (cowfootR)
 
 ## Comprehensive Single Farm Carbon Footprint Analysis
 
-This vignette demonstrates how to conduct a detailed carbon footprint
-assessment for an individual dairy farm using all available functions in
-cowfootR. We’ll work through a realistic example with comprehensive data
-collection and analysis.
+This vignette shows a complete, end-to-end carbon footprint assessment
+for one dairy farm using cowfootR. The goal is to demonstrate how
+individual emission sources can be calculated, combined, visualized, and
+summarized into intensity metrics.
 
 ### Farm Profile: Estancia Las Flores
 
-For this analysis, we’ll assess a medium-sized dairy farm in Uruguay
-with the following characteristics:
+We will analyze a medium-sized dairy farm with a mixed grazing system
+and supplementation:
 
 - **Location**: Temperate climate, well-drained soils
 - **System**: Mixed grazing with supplementation
@@ -20,35 +18,42 @@ with the following characteristics:
 - **Area**: 200 hectares total
 - **Production**: 950,000 litres annually
 
-### Step 1: System Boundaries Definition
+### Step 1: Define system boundaries
 
-First, we define what emission sources to include in our assessment:
+**Purpose**. System boundaries specify which emission sources are
+included in the assessment (e.g., farm-gate vs cradle-to-farm-gate).
+This affects which functions are relevant and how totals are
+interpreted.
 
 ``` r
-# Define comprehensive farm-gate boundaries
+# Most common scope: farm-gate boundaries
 boundaries <- set_system_boundaries("farm_gate")
-print(boundaries)
+boundaries
 #> $scope
 #> [1] "farm_gate"
 #> 
 #> $include
 #> [1] "enteric" "manure"  "soil"    "energy"  "inputs"
+```
 
-# Alternative: cradle-to-farm-gate (includes upstream emissions)
+You can also explore other boundary configurations:
+
+``` r
+# Includes upstream emissions (e.g., purchased inputs)
 boundaries_extended <- set_system_boundaries("cradle_to_farm_gate")
-print(boundaries_extended)
+boundaries_extended
 #> $scope
 #> [1] "cradle_to_farm_gate"
 #> 
 #> $include
 #> [1] "feed"    "enteric" "manure"  "soil"    "energy"  "inputs"
 
-# Custom boundaries example
+# Custom selection of sources
 boundaries_partial <- set_system_boundaries(
   scope = "partial",
   include = c("enteric", "manure", "soil")
 )
-print(boundaries_partial)
+boundaries_partial
 #> $scope
 #> [1] "partial"
 #> 
@@ -56,40 +61,39 @@ print(boundaries_partial)
 #> [1] "enteric" "manure"  "soil"
 ```
 
-For this analysis, we’ll use farm-gate boundaries as they represent the
-most common assessment scope.
+For the rest of this vignette, we use farm-gate boundaries because it is
+a common scope in farm-level reporting.
 
-### Step 2: Detailed Farm Data Collection
+### Step 2: Provide farm data (inputs)
 
-#### Herd Composition and Management
+**Purpose**. cowfootR functions require inputs describing the herd,
[…]
