+++ articles/index.md
@@ -6,6 +6,7 @@
 - [Introduction to Dairy Life Cycle Assessment with
   cowfootR](https://juanmarcosmoreno-arch.github.io/cowfootR/articles/Introduction_to_Dairy_LCA_cowfootR.md):
 - [IPCC_Methodology_Tiers_cowfootR](https://juanmarcosmoreno-arch.github.io/cowfootR/articles/IPCC_Methodology_Tiers_cowfootR.md):
-- [Single_Farm_Analysis_cowfootR](https://juanmarcosmoreno-arch.github.io/cowfootR/articles/Single_Farm_Analysis_cowfootR.md):
+- [Single Farm Analysis
+  (cowfootR)](https://juanmarcosmoreno-arch.github.io/cowfootR/articles/Single_Farm_Analysis_cowfootR.md):
 - [cowfootR
   Workflow](https://juanmarcosmoreno-arch.github.io/cowfootR/articles/workflow.md):
