+++ vignettes/IPCC_Methodology_Tiers_cowfootR.Rmd
@@ -1,16 +1,16 @@
 ---
-title: "IPCC_Methodology_Tiers_cowfootR"
+  title: "IPCC_Methodology_Tiers_cowfootR"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{IPCC_Methodology_Tiers_cowfootR}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
+%\VignetteEngine{knitr::rmarkdown}
+%\VignetteEncoding{UTF-8}
 ---
-
-
-```{r setup, include = FALSE}
+  
+  
+  ```{r setup, include = FALSE}
 knitr::opts_chunk$set(
-  fig.alt = "Figura generada por la viñeta; ver texto para detalles.",
+  fig.alt = "Figure generated by this vignette; see the surrounding text for details.",
   collapse = TRUE,
   comment = "#>",
   fig.width = 8,
@@ -23,38 +23,79 @@ library(cowfootR)
 library(ggplot2)
 library(dplyr)
 library(knitr)
+if (!requireNamespace("tidyr", quietly = TRUE)) {
+  stop("This vignette requires 'tidyr'. Please install it with install.packages('tidyr').")
+}
 ```
 
 # Understanding IPCC Methodology Tiers
 
 The Intergovernmental Panel on Climate Change (IPCC) provides three tiers of methodological complexity for greenhouse gas inventory calculations. cowfootR implements Tier 1 and Tier 2 methodologies for dairy farm carbon footprint assessments. This vignette explains the differences, when to use each approach, and demonstrates their practical application.
 
+## Data Requirements Analysis
+
+### Tier 1 Data Needs
+
+```{r}
+tier1_requirements <- data.frame(
+  Category = c("Animal Data", "Production", "Management", "Optional"),
+  Essential_Data = c(
+    "Number by category (cows, heifers, calves)",
+    "Annual milk production (litres)",
+    "Manure system type, basic inputs",
+    "Farm area, country location"
+  ),
+  Time_to_Collect = c("< 1 hour", "< 1 hour", "1-2 hours", "< 1 hour"),
+  Data_Source = c("Farm records", "Milk processor", "Farmer interview", "Farm records")
+)
+
+kable(tier1_requirements, caption = "Tier 1 Data Requirements")
+```
+
+### Tier 2 Additional Requirements
+
+```{r}
+tier2_additional <- data.frame(
+  Category = c("Animal Characteristics", "Nutrition", "Management Detail", "Environmental"),
+  Additional_Data = c(
+    "Body weights, milk yield per cow, breeding records",
+    "Feed composition, DM intake, diet digestibility",
+    "Precise input quantities, equipment usage",
+    "Climate data, soil types, system temperatures"
+  ),
+  Time_to_Collect = c("2-4 hours", "4-8 hours", "2-4 hours", "1-2 hours"),
+  Expertise_Level = c("Basic", "Intermediate", "Basic", "Basic")
+)
+
+kable(tier2_additional, caption = "Additional Tier 2 Data Requirements")
+```
+
 ## Theoretical Background
 
 ### IPCC Tier System Overview
 
 The IPCC tier system balances accuracy with data requirements and complexity:
-
-- **Tier 1**: Default emission factors with basic activity data
+  
+  - **Tier 1**: Default emission factors with basic activity data
 - **Tier 2**: Country/region-specific factors with detailed activity data  
 - **Tier 3**: Advanced modeling with comprehensive measurements (not implemented in cowfootR)
 
 ### Key Differences in Dairy Applications
 
 | Aspect | Tier 1 | Tier 2 |
-|--------|--------|--------|
[…]
