### 2025-12-30 8b3c1dc
Update IPCC_Methodology_Tiers_cowfootR.Rmd
