### 2026-01-03 #724 @maurolepore
@juanmarcosmoreno-arch similar a las revisiones académicas, por favor responde  a los comentarios de revisión lo más explícitamente posible. La intención es que las personas que revisaron el paquete puedan ver claramente tu respuesta a cada una de sus sugerencias.

Veo que Paola uso el mismo formato que yo, numerando sus comentarios y clasificados  como pedidos y opcionales. Se esperaría respuesta a todos los pedidos no-opcionales, incluyendo la etiqueta pc01, pc02, etc.

El formato está explicado acá:
https://github.com/ropensci/software-review/issues/724#issuecomment-3477950381

Mi explicación fue un poco tersa. Avísame si no está claro.
