### 2026-01-05 #724 @juanmarcosmoreno-arch
Muchas gracias @maurolepore por la revisión y por las sugerencias concretas! A continuación respondo punto por punto y detallo los cambios que ya realicé.

ml02
Se reorganizó la sección Reference del sitio pkgdown, agrupando las funciones por categoría.

ml03
Se simplificó el README para incluir un único ejemplo mínimo que muestra el flujo principal del paquete. Los ejemplos extendidos y el uso detallado fueron movidos a una viñeta “Get started with cowfootR”

ml04
Se actualizaron los ejemplos del README para mostrar explícitamente el output de las funciones, sin el uso innecesario de print()

ml05.
Se separaron claramente las instrucciones de instalación desde CRAN y desde GitHub en bloques de código independientes, facilitando el copiado y ejecución directa por parte de los usuarios.

gracias
