+++ README.Rmd
@@ -27,7 +27,12 @@ and IPCC guidelines for greenhouse gas accounting.
 
 ## Overview
 
-`cowfootR` provides a comprehensive toolkit for calculating carbon footprints of dairy farms following IPCC guidelines. The package includes:
+`cowfootR` provides a comprehensive toolkit for calculating carbon
+footprints of dairy farms following IPCC guidelines
+([IPCC 2019 Refinement](https://www.ipcc-nggip.iges.or.jp/public/2019rf/index.html))
+and International Dairy Federation guidance for the dairy sector
+([IDF Bulletin 520](https://shop.fil-idf.org/products/the-idf-global-carbon-footprint-standard-for-the-dairy-sector?_pos=1&_sid=8a3f414f8&_ss=r)).
+The package includes:
 
 - **Individual emission calculations** from enteric fermentation, manure, soil, energy, and inputs
 - **Batch processing** capabilities for multiple farms
@@ -37,21 +42,21 @@ and IPCC guidelines for greenhouse gas accounting.
 
 ## Installation
 
-You can install the development version of cowfootR from GitHub:
 
-```{r installation, eval=FALSE}
-# Install development version from GitHub
-# install.packages("devtools")
+```{r eval=FALSE}
+install.packages("cowfootR")
+```
+Or install the development version:
+```{r eval=FALSE}
 devtools::install_github("juanmarcosmoreno-arch/cowfootR")
 ```
 
 ## Quick Start
 
-### Single Farm Analysis
+Below is a minimal, end-to-end example showing the core workflow of `cowfootR`
+for a single dairy farm.
 
-This is a basic example of how to calculate the carbon footprint of a single farm:
-
-```{r example, eval=FALSE}
+```{r example}
 library(cowfootR)
 
 # 1. Define system boundaries
@@ -59,115 +64,55 @@ boundaries <- set_system_boundaries("farm_gate")
 
 # 2. Calculate emissions by source
 enteric <- calc_emissions_enteric(
-  n_animals = 100, 
+  n_animals = 100,
   cattle_category = "dairy_cows",
   boundaries = boundaries
 )
 
 manure <- calc_emissions_manure(
-  n_cows = 100, 
+  n_cows = 100,
   boundaries = boundaries
 )
 
 soil <- calc_emissions_soil(
-  n_fertilizer_synthetic = 1500, 
-  n_excreta_pasture = 5000, 
+  n_fertilizer_synthetic = 1500,
+  n_excreta_pasture = 5000,
   area_ha = 120,
   boundaries = boundaries
 )
 
 energy <- calc_emissions_energy(
-  diesel_l = 2000, 
-  electricity_kwh = 5000, 
+  diesel_l = 2000,
+  electricity_kwh = 5000,
   boundaries = boundaries
 )
 
 inputs <- calc_emissions_inputs(
-  conc_kg = 1000, 
-  fert_n_kg = 500, 
+  conc_kg = 1000,
+  fert_n_kg = 500,
   boundaries = boundaries
 )
 
-# 3. Calculate total emissions
+# 3. Aggregate total emissions
 total_emissions <- calc_total_emissions(enteric, manure, soil, energy, inputs)
-print(paste("Total emissions:", round(total_emissions$total_co2eq, 1), "kg CO2eq"))
+total_emissions
 
-# 4. Calculate intensity metrics
+# 4. Intensity metrics
 milk_intensity <- calc_intensity_litre(
   total_emissions = total_emissions,
   milk_litres = 750000,
   fat = 4.0,
   protein = 3.3
 )
-print(paste("Milk intensity:", round(milk_intensity$intensity_co2eq_per_kg_fpcm, 2), 
-            "kg CO2eq/kg FPCM"))
[…]
