+++ README.md
@@ -1,4 +1,5 @@
 
+
 # cowfootR <img src="man/figures/logo.png" align="right" width="120" />
 
 Tools to estimate the carbon footprint of dairy farms.  
@@ -37,20 +38,20 @@ The package includes:
 
 ## Installation
 
-You can install the development version of cowfootR from GitHub:
+``` r
+install.packages("cowfootR")
+```
+
+Or install the development version:
 
 ``` r
-# Install development version from GitHub
-# install.packages("devtools")
 devtools::install_github("juanmarcosmoreno-arch/cowfootR")
 ```
 
 ## Quick Start
 
-### Single Farm Analysis
-
-This is a basic example of how to calculate the carbon footprint of a
-single farm:
+Below is a minimal, end-to-end example showing the core workflow of
+`cowfootR` for a single dairy farm.
 
 ``` r
 library(cowfootR)
@@ -60,115 +61,92 @@ boundaries <- set_system_boundaries("farm_gate")
 
 # 2. Calculate emissions by source
 enteric <- calc_emissions_enteric(
-  n_animals = 100, 
+  n_animals = 100,
   cattle_category = "dairy_cows",
   boundaries = boundaries
 )
 
 manure <- calc_emissions_manure(
-  n_cows = 100, 
+  n_cows = 100,
   boundaries = boundaries
 )
 
 soil <- calc_emissions_soil(
-  n_fertilizer_synthetic = 1500, 
-  n_excreta_pasture = 5000, 
+  n_fertilizer_synthetic = 1500,
+  n_excreta_pasture = 5000,
   area_ha = 120,
   boundaries = boundaries
 )
 
 energy <- calc_emissions_energy(
-  diesel_l = 2000, 
-  electricity_kwh = 5000, 
+  diesel_l = 2000,
+  electricity_kwh = 5000,
   boundaries = boundaries
 )
 
 inputs <- calc_emissions_inputs(
-  conc_kg = 1000, 
-  fert_n_kg = 500, 
+  conc_kg = 1000,
+  fert_n_kg = 500,
   boundaries = boundaries
 )
 
-# 3. Calculate total emissions
+# 3. Aggregate total emissions
 total_emissions <- calc_total_emissions(enteric, manure, soil, energy, inputs)
-print(paste("Total emissions:", round(total_emissions$total_co2eq, 1), "kg CO2eq"))
+total_emissions
+#> Carbon Footprint - Total Emissions
+#> ==================================
+#> Total CO2eq: 451512.6 kg
+#> Number of sources: 5 
+#> 
+#> Breakdown by source:
+#>   energy : 5740 kg CO2eq
+#>   enteric : 312800 kg CO2eq
+#>   inputs : 4000 kg CO2eq
+#>   manure : 89880 kg CO2eq
+#>   soil : 39092.62 kg CO2eq
+#> 
+#> Calculated on: 2026-01-05
 
-# 4. Calculate intensity metrics
+# 4. Intensity metrics
 milk_intensity <- calc_intensity_litre(
   total_emissions = total_emissions,
   milk_litres = 750000,
   fat = 4.0,
   protein = 3.3
 )
-print(paste("Milk intensity:", round(milk_intensity$intensity_co2eq_per_kg_fpcm, 2), 
-            "kg CO2eq/kg FPCM"))
+milk_intensity
+#> Carbon Footprint Intensity
+#> ==========================
+#> Intensity: 0.585 kg CO2eq/kg FPCM
+#> 
+#> Production data:
+#>  Raw milk (L): 750,000 L
+#>  Raw milk (kg): 772,500 kg
+#>  FPCM (kg): 772,407 kg
[…]
