### 2026-01-05 78c65b6
Update test-calc_batch.R
