+++ tests/testthat/test-calc_batch.R
@@ -1,4 +1,4 @@
-# Tests for batch processing
+# Tests for batch processing -----------------------------------------------
 
 test_that("calc_batch processes multiple farms", {
   farms <- data.frame(
@@ -21,10 +21,23 @@ test_that("calc_batch processes multiple farms", {
 })
 
 test_that("calc_batch validates tier input", {
-  farms <- data.frame(FarmID = "A", Milk_litres = 500000)
+  farms <- data.frame(
+    FarmID = "A",
+    Milk_litres = 500000,
+    stringsAsFactors = FALSE
+  )
 
-  expect_error(calc_batch(data = farms, tier = 3))
-  expect_error(calc_batch(data = data.frame()))
+  expect_error(
+    calc_batch(data = farms, tier = 3),
+    regexp = "tier.*(1|2)|invalid.*tier|must be.*(1|2)",
+    ignore.case = TRUE
+  )
+
+  expect_error(
+    calc_batch(data = data.frame()),
+    regexp = "data.*(empty|no rows)|nrow\\(data\\).*0|must contain",
+    ignore.case = TRUE
+  )
 })
 
 test_that("calc_batch handles errors gracefully", {
@@ -35,7 +48,13 @@ test_that("calc_batch handles errors gracefully", {
     stringsAsFactors = FALSE
   )
 
-  result <- suppressMessages(calc_batch(data = farms, tier = 1))
+  result <- suppressMessages(
+    calc_batch(
+      data = farms,
+      tier = 1,
+      boundaries = set_system_boundaries("farm_gate")
+    )
+  )
 
   expect_equal(result$summary$n_farms_successful, 1)
   expect_equal(result$summary$n_farms_with_errors, 1)
