### 2026-01-05 5dd7bfd
Update test-calc_emissions_energy.R
