### 2026-01-05 de60a5f
Update test-calc_emissions_enteric.R
