### 2026-01-05 33a3eaf
Update test-calc_emissions_inputs.R
