### 2026-01-05 cd8a96f
Update test-calc_emissions_manure.R
