### 2026-01-05 a82a81d
Update test-calc_intensity.R
