### 2026-01-05 75ca3f4
Update test-calc_total_emissions.R
