### 2026-01-05 c9bfac0
Update test-energy.R
