+++ tests/testthat/test-intensities.R
@@ -1,148 +1,91 @@
-testthat::test_that("setup: intensities functions exist", {
-  testthat::skip_if_not(exists("calc_intensity_area"))
-  testthat::skip_if_not(exists("calc_intensity_litre"))
+# Tests for intensity calculations (strong + consistent)
+
+test_that("calc_intensity_litre returns an S3 cf_intensity with finite positive intensity", {
+  res <- calc_intensity_litre(
+    total_emissions = 85000,
+    milk_litres = 750000,
+    fat = 3.9,
+    protein = 3.2
+  )
+
+  expect_s3_class(res, "cf_intensity")
+  expect_true(is.numeric(res$intensity_co2eq_per_kg_fpcm))
+  expect_true(is.finite(res$intensity_co2eq_per_kg_fpcm))
+  expect_gt(res$intensity_co2eq_per_kg_fpcm, 0)
+
+  expect_equal(res$milk_production_litres, 750000)
 })
 
-# -------------------------------------------------------------------------
-# calc_intensity_area tests
-# -------------------------------------------------------------------------
+test_that("calc_intensity_litre decreases when milk increases (same emissions)", {
+  res_lo <- calc_intensity_litre(
+    total_emissions = 85000,
+    milk_litres = 700000,
+    fat = 3.9,
+    protein = 3.2
+  )
 
-testthat::test_that("calc_intensity_area: basic case returns finite non-negative intensity", {
-  fn <- calc_intensity_area
+  res_hi <- calc_intensity_litre(
+    total_emissions = 85000,
+    milk_litres = 800000,
+    fat = 3.9,
+    protein = 3.2
+  )
 
-  # Emissions total and area values for test
-  total <- 1000
-  area  <- 50
-
-  # Try different calling styles depending on function signature
-  out <- try(safe_call(fn,
-                       canonical_args = list(total_CO2eq = total, area_ha = area),
-                       positional_args = list(total, area),
-                       df_args = list(total_CO2eq = total, area_ha = area)
-  ), silent = TRUE)
-
-  if (inherits(out, "try-error")) testthat::skip("Function signature incompatible — skipping test.")
-
-  val <- pick_named_numeric(out, patterns = c("per_ha", "intens", "ha"))
-  if (length(val) == 0) testthat::skip("No intensity field detected in output.")
-
-  testthat::expect_true(all(is.finite(val)))
-  testthat::expect_true(all(val >= 0))
+  expect_lt(res_hi$intensity_co2eq_per_kg_fpcm, res_lo$intensity_co2eq_per_kg_fpcm)
 })
 
-testthat::test_that("calc_intensity_area: higher emissions with same area → higher intensity", {
-  fn <- calc_intensity_area
+test_that("calc_intensity_litre validates inputs (specific failures)", {
+  expect_error(
+    calc_intensity_litre(total_emissions = -1, milk_litres = 750000, fat = 4, protein = 3.3),
+    regexp = "total|emission|positive|>=\\s*0",
+    ignore.case = TRUE
+  )
 
-  o1 <- try(safe_call(fn,
-                      canonical_args = list(total_CO2eq = 1000, area_ha = 100),
-                      positional_args = list(1000, 100),
-                      df_args = list(total_CO2eq = 1000, area_ha = 100)
-  ), silent = TRUE)
-
-  o2 <- try(safe_call(fn,
-                      canonical_args = list(total_CO2eq = 1100, area_ha = 100),
-                      positional_args = list(1100, 100),
-                      df_args = list(total_CO2eq = 1100, area_ha = 100)
-  ), silent = TRUE)
-
[…]
