+++ tests/testthat/test-manure.R
@@ -1,84 +1,94 @@
-testthat::test_that("setup: manure function exists", {
-  testthat::skip_if_not(exists("calc_emissions_manure"))
-})
+# Tests for manure management emissions (API-aligned, strong expectations)
 
-# -------------------------------------------------------------------------
-# calc_emissions_manure tests
-# -------------------------------------------------------------------------
-
-testthat::test_that("manure: minimal structure contains expected components", {
-  fn <- calc_emissions_manure
-
-  out <- try(safe_call(fn,
-                       canonical_args = list(storage = "lagoon", digestor_eff = 0.0, n_excreted_kg = 50, days_storage = 90),
-                       positional_args = list("lagoon", 0.0, 50, 90),
-                       df_args = list(storage = "lagoon", digestor_eff = 0.0, n_excreted_kg = 50, days_storage = 90)
-  ), silent = TRUE)
-
-  if (inherits(out, "try-error")) testthat::skip("Function signature incompatible — skipping test.")
-
-  testthat::expect_true(is.list(out) || is.data.frame(out))
-  nm <- names(out)
-  testthat::expect_true(any(grepl("CH4",   nm, ignore.case = TRUE)))
-  testthat::expect_true(any(grepl("N2O",   nm, ignore.case = TRUE)))
-  testthat::expect_true(any(grepl("total", nm, ignore.case = TRUE)))
-})
-
-testthat::test_that("manure: different storage types and digester reduce emissions", {
-  fn <- calc_emissions_manure
-
-  storages <- c("lagoon", "solid_stack", "compost")
-  res <- lapply(storages, function(s) {
-    try(safe_call(fn,
-                  canonical_args = list(storage = s, digestor_eff = 0.0, n_excreted_kg = 50, days_storage = 90),
-                  positional_args = list(s, 0.0, 50, 90),
-                  df_args = list(storage = s, digestor_eff = 0.0, n_excreted_kg = 50, days_storage = 90)
-    ), silent = TRUE)
-  })
-
-  if (any(vapply(res, inherits, logical(1), "try-error"))) testthat::skip("Signature incompatible — skipping storage test.")
-
-  totals <- vapply(res, function(x) pick_named_numeric(x, patterns = c("total"))[1], numeric(1))
-  testthat::expect_true(all(is.finite(totals)))
-
-  sin_dig <- safe_call(fn,
-                       canonical_args = list(storage = "lagoon", digestor_eff = 0.0, n_excreted_kg = 50, days_storage = 90),
-                       positional_args = list("lagoon", 0.0, 50, 90),
-                       df_args = list(storage = "lagoon", digestor_eff = 0.0, n_excreted_kg = 50, days_storage = 90)
-  )
-  con_dig <- safe_call(fn,
-                       canonical_args = list(storage = "lagoon", digestor_eff = 0.7, n_excreted_kg = 50, days_storage = 90),
-                       positional_args = list("lagoon", 0.7, 50, 90),
-                       df_args = list(storage = "lagoon", digestor_eff = 0.7, n_excreted_kg = 50, days_storage = 90)
+test_that("calc_emissions_manure returns expected basic structure (Tier 1)", {
+  res <- calc_emissions_manure(
+    n_cows = 100,
+    manure_system = "pasture",
+    tier = 1
   )
 
-  t0 <- pick_named_numeric(sin_dig, patterns = c("total"))[1]
-  t1 <- pick_named_numeric(con_dig, patterns = c("total"))[1]
[…]
