+++ tests/testthat/test-total.R
@@ -1,116 +1,52 @@
-testthat::test_that("setup: total function exists", {
-  testthat::skip_if_not(exists("calc_total_emissions"))
+# Tests for total emissions aggregation (calc_total_emissions)
+
+test_that("calc_total_emissions aggregates correctly with standard fields", {
+  e1 <- list(source = "enteric", co2eq_kg = 1000)
+  e2 <- list(source = "manure",  co2eq_kg = 500)
+  e3 <- list(source = "soil",    co2eq_kg = 300)
+
+  result <- calc_total_emissions(e1, e2, e3)
+
+  expect_s3_class(result, "cf_total")
+  expect_equal(result$total_co2eq, 1800)
+  expect_equal(result$n_sources, 3)
 })
 
-# -------------------------------------------------------------------------
-# calc_total_emissions tests
-# -------------------------------------------------------------------------
-# These tests do not assume a fixed input signature.
-# Strategy:
-#   1) Try to pass a compact list/df of component totals.
-#   2) If incompatible, skip gracefully.
-# Assertions:
-#   - Basic case returns a finite/non-negative total-like value.
-#   - Monotonicity: increasing one component increases the final total.
-#   - Optional: if component fields are detectable in the output, check
-#     that "reported total" matches the sum of components within tolerance.
+test_that("calc_total_emissions supports alternative CO2eq field names", {
+  e1 <- list(source = "enteric", co2eq_kg = 1000)
+  e2 <- list(source = "manure",  total_co2eq_kg = 500)
+  e3 <- list(source = "soil",    total_co2eq = 300)
 
-# Helper to build a component set (names are common but may vary in your impl)
-.build_components <- function(k_enteric = 200, k_manure = 100, k_soil = 50, k_energy = 80, k_inputs = 30) {
-  list(
-    enteric = k_enteric,
-    manure  = k_manure,
-    soil    = k_soil,
-    energy  = k_energy,
-    inputs  = k_inputs
+  result <- calc_total_emissions(e1, e2, e3)
+
+  expect_s3_class(result, "cf_total")
+  expect_equal(result$total_co2eq, 1800)
+})
+
+test_that("calc_total_emissions is monotonic (increasing one source increases total)", {
+  base <- calc_total_emissions(
+    list(source = "enteric", co2eq_kg = 1000),
+    list(source = "manure",  co2eq_kg = 500)
   )
-}
 
-testthat::test_that("total: basic case returns finite/non-negative total", {
-  fn <- calc_total_emissions
-  comps <- .build_components()
+  more <- calc_total_emissions(
+    list(source = "enteric", co2eq_kg = 1100),
+    list(source = "manure",  co2eq_kg = 500)
+  )
 
-  out <- try(safe_call(fn,
-                       canonical_args = comps,
-                       positional_args = unname(comps),
-                       df_args = comps
-  ), silent = TRUE)
-
-  if (inherits(out, "try-error")) testthat::skip("Incompatible signature — skipping total basic.")
-
-  # Try to pick a 'total' field; if none, use any finite numeric as fallback
-  tot <- pick_named_numeric(out, patterns = c("total", "co2", "emiss"))
-  if (length(tot) == 0) tot <- pluck_numeric(out)
-
-  testthat::expect_true(length(tot) >= 1)
-  testthat::expect_true(all(is.finite(tot)))
-  testthat::expect_true(all(tot >= 0))
+  expect_gt(more$total_co2eq, base$total_co2eq)
 })
 
-testthat::test_that("total: increasing one component increases final total", {
-  fn <- calc_total_emissions
[…]
