### 2026-01-08 8bd6a3b
Update paper.md
