### 2026-01-08 942b513
Docs: clarify units and boundaries; add calc_batch workflow example
