+++ R/calc_emissions_soil.R
@@ -40,8 +40,8 @@
 #' calc_emissions_soil(n_fertilizer_synthetic = 2000, include_indirect = FALSE)
 #'
 #' # Boundary exclusion example
-#' b <- list(include = c("energy", "manure"))  # soil not included
-#' calc_emissions_soil(n_fertilizer_synthetic = 1000, boundaries = b)$co2eq_kg  # 0
+#' b <- list(include = c("energy", "manure")) # soil not included
+#' calc_emissions_soil(n_fertilizer_synthetic = 1000, boundaries = b)$co2eq_kg # 0
 #' }
 calc_emissions_soil <- function(n_fertilizer_synthetic = 0,
                                 n_fertilizer_organic = 0,
@@ -54,34 +54,37 @@ calc_emissions_soil <- function(n_fertilizer_synthetic = 0,
                                 include_indirect = TRUE,
                                 gwp_n2o = 273,
                                 boundaries = NULL) {
-
   # --------------------------
   # Validation & boundary gate
   # --------------------------
-  valid_soils    <- c("well_drained", "poorly_drained")
+  valid_soils <- c("well_drained", "poorly_drained")
   valid_climates <- c("temperate", "tropical")
 
-  if (!soil_type %in% valid_soils)
+  if (!soil_type %in% valid_soils) {
     stop("Invalid soil_type. Use: ", paste(valid_soils, collapse = ", "))
-  if (!climate %in% valid_climates)
+  }
+  if (!climate %in% valid_climates) {
     stop("Invalid climate. Use: ", paste(valid_climates, collapse = ", "))
+  }
 
   nin <- c(n_fertilizer_synthetic, n_fertilizer_organic, n_excreta_pasture, n_crop_residues)
-  if (any(!is.finite(nin)))
+  if (any(!is.finite(nin))) {
     stop("Nitrogen inputs must be finite numeric scalars.")
-  if (any(nin < 0))
+  }
+  if (any(nin < 0)) {
     stop("All nitrogen inputs must be non-negative.")
+  }
 
   # Boundary exclusion: return numeric zero (not NULL) for co2eq_kg as per tests.
   soil_excluded <- is.list(boundaries) && !is.null(boundaries$include) && !("soil" %in% boundaries$include)
   if (soil_excluded) {
     return(list(
-      source             = "soil",
-      co2eq_kg           = 0,  # numeric zero required by tests
-      methodology        = "excluded_by_boundaries",
+      source = "soil",
+      co2eq_kg = 0, # numeric zero required by tests
+      methodology = "excluded_by_boundaries",
       emissions_breakdown = NULL,
-      nitrogen_inputs     = NULL,
-      date               = Sys.Date()
+      nitrogen_inputs = NULL,
+      date = Sys.Date()
     ))
   }
 
@@ -99,10 +102,12 @@ calc_emissions_soil <- function(n_fertilizer_synthetic = 0,
     )
     ef_direct <- direct_factors[[climate]][[soil_type]]
   } else {
-    if (!is.numeric(ef_direct) || length(ef_direct) != 1L)
+    if (!is.numeric(ef_direct) || length(ef_direct) != 1L) {
       stop("ef_direct must be a single numeric value.")
-    if (ef_direct < 0 || ef_direct > 0.05)
+    }
+    if (ef_direct < 0 || ef_direct > 0.05) {
       warning("ef_direct seems unusual (typical range is ~0.005 to 0.02).")
+    }
   }
 
   # -------------------------
[…]
