+++ R/set_system_boundaries.R
@@ -12,9 +12,10 @@
 #' b1 <- set_system_boundaries("farm_gate")
 #' b2 <- set_system_boundaries(include = c("enteric", "manure", "soil"))
 #' b3 <- set_system_boundaries(include = c("enteric", "manure"))
-#' b1$scope; b2$include; b3$include
+#' b1$scope
+#' b2$include
+#' b3$include
 set_system_boundaries <- function(scope = "farm_gate", include = NULL) {
-
   scope <- match.arg(scope, c("farm_gate", "cradle_to_farm_gate", "partial"))
 
   if (scope == "farm_gate") {
@@ -33,4 +34,3 @@ set_system_boundaries <- function(scope = "farm_gate", include = NULL) {
     include = include
   )
 }
-

[…]
