### 2026-01-08 9de595d
Deploying to gh-pages from @ juanmarcosmoreno-arch/cowfootR@942b5133bcb145a018312e46d41c26f6e3799254 🚀
