+++ news/index.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## cowfootR 0.1.2
+
+CRAN release: 2025-09-22
+
 ## cowfootR 0.1.1
 
 ## cowfootR 0.1.0
