+++ articles/index.md
@@ -1,12 +1,21 @@
 # Articles
 
-### All vignettes
+### Get started
 
-- [Complete_Parameter_Reference_Guide](https://juanmarcosmoreno-arch.github.io/cowfootR/articles/Complete_Parameter_Reference_Guide.md):
+- [Get started with
+  cowfootR](https://juanmarcosmoreno-arch.github.io/cowfootR/articles/00%20-%20Get%20started%20with%20cowfootR.md):
+
+### Core workflows
+
 - [Introduction to Dairy Life Cycle Assessment with
-  cowfootR](https://juanmarcosmoreno-arch.github.io/cowfootR/articles/Introduction_to_Dairy_LCA_cowfootR.md):
-- [IPCC_Methodology_Tiers_cowfootR](https://juanmarcosmoreno-arch.github.io/cowfootR/articles/IPCC_Methodology_Tiers_cowfootR.md):
+  cowfootR](https://juanmarcosmoreno-arch.github.io/cowfootR/articles/01%20-%20Introduction%20to%20Dairy%20Life%20Cycle%20Assessment.md):
 - [Single Farm Analysis
-  (cowfootR)](https://juanmarcosmoreno-arch.github.io/cowfootR/articles/Single_Farm_Analysis_cowfootR.md):
+  (cowfootR)](https://juanmarcosmoreno-arch.github.io/cowfootR/articles/02%20-%20Single%20Farm%20Analysis%20(cowfootR).md):
+- [IPCC Methodology Tiers in
+  cowfootR](https://juanmarcosmoreno-arch.github.io/cowfootR/articles/03%20-%20IPCC%20Methodology%20Tiers%20cowfootR.md):
+
+### Reference
+
+- [Complete_Parameter_Reference_Guide](https://juanmarcosmoreno-arch.github.io/cowfootR/articles/04%20-%20Complete%20Parameter%20Reference%20Guide.md):
 - [cowfootR
   Workflow](https://juanmarcosmoreno-arch.github.io/cowfootR/articles/workflow.md):
