+++ authors.md
@@ -9,14 +9,14 @@
 Source:
 [`DESCRIPTION`](https://github.com/juanmarcosmoreno-arch/cowfootR/blob/main/DESCRIPTION)
 
-Moreno J (2025). *cowfootR: Dairy Farm Carbon Footprint Assessment*. R
-package version 0.1.1,
+Moreno J (2026). *cowfootR: Dairy Farm Carbon Footprint Assessment*. R
+package version 0.1.2,
 <https://github.com/juanmarcosmoreno-arch/cowfootR>.
 
     @Manual{,
       title = {cowfootR: Dairy Farm Carbon Footprint Assessment},
       author = {Juan Moreno},
-      year = {2025},
-      note = {R package version 0.1.1},
+      year = {2026},
+      note = {R package version 0.1.2},
       url = {https://github.com/juanmarcosmoreno-arch/cowfootR},
     }
