### 2026-01-08 3f1e63b
Fix tests for system boundaries and total emissions aggregation
