### 2026-01-08 1b77ee1
Deploying to gh-pages from @ juanmarcosmoreno-arch/cowfootR@3f1e63b0b553b58346668d672b088650203fd172 🚀
