+++ pkgdown.yml
@@ -9,7 +9,7 @@ articles:
   03 - IPCC Methodology Tiers cowfootR: 03 - IPCC Methodology Tiers cowfootR.html
   04 - Complete Parameter Reference Guide: 04 - Complete Parameter Reference Guide.html
   workflow: workflow.html
-last_built: 2026-01-08T12:17Z
+last_built: 2026-01-08T13:06Z
 urls:
   reference: https://juanmarcosmoreno-arch.github.io/cowfootR/reference
   article: https://juanmarcosmoreno-arch.github.io/cowfootR/articles
