### 2026-01-08 476856d
Update paper.md
