### 2026-01-08 dd0c5d5
Deploying to gh-pages from @ juanmarcosmoreno-arch/cowfootR@476856d2422f2dc79c7f0bbd81a14db3bb64ccae 🚀
