### 2026-01-08 f2424f1
Update paper.bib
