### 2026-01-08 377d319
Deploying to gh-pages from @ juanmarcosmoreno-arch/cowfootR@f2424f134adf164c52149e51f9fe11e2e826daf2 🚀
