### 2026-01-12 ab50104
Deploying to gh-pages from @ juanmarcosmoreno-arch/cowfootR@d4217a50b9e78f783001340b1e0f560bb91496f6 🚀
