### 2026-01-12 ab1c1fe
Fix tests for system boundaries and total emissions aggregation
