+++ R/calc_batch.R
@@ -126,12 +126,49 @@ compute_fpcm_idf <- function(milk_l, fat, prot, milk_density = 1.03) {
 #'
 #' @details
 #' The input data frame is intentionally flexible to support heterogeneous
-#' data availability across farms. Column names follow cowfootR conventions
-#' (see vignettes for a complete template). Missing optional columns do not
-#' cause errors; instead, tier-specific default assumptions are applied.
+#' data availability across farms. Each row represents one farm and all inputs
+#' are assumed to correspond to a single accounting year, unless explicitly
+#' stated otherwise by the column name (e.g., *_kg_day).
 #'
-#' A minimal valid input contains farm identification, milk production, and
-#' herd size, while extended inputs enable refined emission estimates.
+#' Column names follow cowfootR conventions. The complete and authoritative
+#' specification of supported input columns (including expected units and
+#' whether they are required or optional) is provided by the Excel template
+#' generated with \code{cf_download_template()}. This template represents the
+#' full set of columns that can be used by \code{calc_batch()}.
+#'
+#' In addition, the vignettes "Get started" and "IPCC Methodology Tiers in cowfootR"
+#' describe how these columns are used conceptually and methodologically.
+#'
+#' \strong{Tier 2–relevant optional columns:}
+#' When \code{tier = 2}, \code{calc_batch()} uses additional farm-specific
+#' information if available. The most relevant optional columns include:
+#' \itemize{
+#'   \item \strong{Enteric fermentation:}
+#'   \code{Milk_yield_kg_cow_year},
+#'   \code{Body_weight_cows_kg},
+#'   \code{MS_intake_cows_milking_kg_day},
+#'   \code{Ym_percent}.
+#'
+#'   \item \strong{Young stock (optional refinement):}
+#'   \code{Body_weight_heifers_kg},
+#'   \code{Body_weight_calves_kg},
+#'   \code{Body_weight_bulls_kg},
+#'   \code{MS_intake_heifers_kg_day},
+#'   \code{MS_intake_calves_kg_day},
+#'   \code{MS_intake_bulls_kg_day}.
+#'
+#'   \item \strong{Manure management:}
+#'   \code{Manure_system},
+#'   \code{Diet_digestibility},
+#'   \code{Protein_intake_kg_day},
+#'   \code{Retention_days},
+#'   \code{System_temperature},
+#'   \code{Climate_zone}.
+#' }
+#'
+#' If any Tier 2–relevant column is missing, the function automatically falls
+#' back to Tier 1–consistent default assumptions following IPCC and IDF guidance.
+#' Missing optional inputs therefore do not cause errors.
 #'
 #' @return A list with \code{$summary} and \code{$farm_results}; class \code{cf_batch_complete}.
 #' Absolute emissions returned in \code{$farm_results} (e.g., \code{emissions_total},
