### 2026-01-12 5ab82d7
Deploying to gh-pages from @ juanmarcosmoreno-arch/cowfootR@ab1c1fe40cb34f4b7ba1adf735071feaede410b7 🚀
