+++ reference/export_hdc_report.md
@@ -70,7 +70,7 @@ class(br) <- "cf_batch_complete"
 
 f <- tempfile(fileext = ".xlsx")
 export_hdc_report(br, file = f)
-#> Batch report saved to: /tmp/Rtmp6slZLD/file1b5267d5603d.xlsx
+#> Batch report saved to: /tmp/Rtmpvavx4d/file1b5249275588.xlsx
 file.exists(f)
 #> [1] TRUE
 # }
