### 2026-01-12 2692b69
Update 03 - IPCC Methodology Tiers cowfootR.Rmd
