### 2026-01-12 ea56209
Deploying to gh-pages from @ juanmarcosmoreno-arch/cowfootR@2692b6953a460db9487c162e8673bcce24d6ce95 🚀
