### 2026-01-13 b1d0fee
Sync: update units + tests + docs
