+++ R/calc_emissions_inputs.R
@@ -65,6 +65,22 @@ calc_emissions_inputs <- function(conc_kg = 0,
                                   ef_fert = NULL,
                                   ef_plastic = NULL,
                                   boundaries = NULL) {
+
+  # ---- Units (annual) -------------------------------------------------------
+  units <- list(
+    co2eq_kg = "kg CO2eq yr-1",
+    conc_kg = "kg yr-1",
+    fert_n_kg = "kg N yr-1",
+    plastic_kg = "kg yr-1",
+    feed_kg = "kg DM yr-1",
+    transport_km = "km",
+    ef_conc = "kg CO2e per kg",
+    ef_fert = "kg CO2e per kg N",
+    ef_plastic = "kg CO2e per kg",
+    ef_feed = "kg CO2e per kg DM",
+    ef_truck = "kg CO2e per (kg·km)"
+  )
+
   # -------------------------------
   # Input validation and boundaries
   # -------------------------------
@@ -78,6 +94,7 @@ calc_emissions_inputs <- function(conc_kg = 0,
   if (!fert_type %in% valid_fert_types) {
     stop("Invalid `fert_type`. Use one of: ", paste(valid_fert_types, collapse = ", "))
   }
+
   # Normalize plastic type case-safely
   if (tolower(plastic_type) == "mixed") {
     plastic_type <- "mixed"
@@ -88,20 +105,23 @@ calc_emissions_inputs <- function(conc_kg = 0,
     }
   }
 
-  # Non-negative quantities check
+  # Non-negative & finite quantities check
   qty <- c(
     conc_kg, fert_n_kg, plastic_kg, feed_grain_dry_kg, feed_grain_wet_kg,
     feed_ration_kg, feed_byproducts_kg, feed_proteins_kg,
     feed_corn_kg, feed_soy_kg, feed_wheat_kg
   )
+  if (any(!is.finite(qty))) stop("All input quantities must be finite numeric scalars.")
   if (any(qty < 0, na.rm = TRUE)) stop("All input quantities must be non-negative.")
 
   # Boundary exclusion: return numeric zero when "inputs" not included
-  if (!is.null(boundaries) && !("inputs" %in% boundaries$include)) {
+  if (is.list(boundaries) && !is.null(boundaries$include) &&
+      !("inputs" %in% boundaries$include)) {
     return(list(
       source = "inputs",
       co2eq_kg = 0,
       total_co2eq_kg = 0,
+      units = units,
       methodology = "excluded_by_boundaries",
       emissions_breakdown = NULL,
       inputs_summary = NULL,
@@ -159,17 +179,14 @@ calc_emissions_inputs <- function(conc_kg = 0,
   # -------------------------------
   to_num_safe <- function(x) {
     y <- suppressWarnings(as.numeric(x))
-    if (is.null(x) || length(y) == 0L || !is.finite(y)) {
-      return(0)
-    }
+    if (is.null(x) || length(y) == 0L || !is.finite(y)) return(0)
     y
   }
   transport_km <- to_num_safe(transport_km)
 
   transport_adjustment <- 0
+  ef_truck <- 1e-4  # kg CO2e per (kg·km)
   if (transport_km > 0) {
-    # Default LCA truck factor (kg CO2e per kg·km); keep conservative order of magnitude
-    ef_truck <- 1e-4
     total_feed_kg <- sum(
       c(
         conc_kg, feed_grain_dry_kg, feed_grain_wet_kg, feed_ration_kg,
@@ -237,6 +254,8 @@ calc_emissions_inputs <- function(conc_kg = 0,
   # -------------------------------
   result <- list(
[…]
