### 2026-01-13 9e2f8ad
Deploying to gh-pages from @ juanmarcosmoreno-arch/cowfootR@db386985d630b39c643ecea0c96c2c6a0958bc76 🚀
