+++ reference/print.cf_intensity.md
@@ -36,7 +36,7 @@ x <- list(
   total_emissions_co2eq = 85000,
   milk_production_litres = 750000,
   milk_production_kg = 750000 * 1.03,
-  fpcm_production_kg = 750000 * 1.03 * (0.1226*4 + 0.0776*3.3 + 0.2534),
+  fpcm_production_kg = 750000 * 1.03 * (0.1226 * 4 + 0.0776 * 3.3 + 0.2534),
   fat_percent = 4, protein_percent = 3.3, milk_density_kg_per_l = 1.03,
   date = Sys.Date()
 )
