+++ NEWS.md
@@ -1,3 +1,5 @@
+# cowfootR 0.1.3
+
 # cowfootR 0.1.2
 
 # cowfootR 0.1.1
