+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: cowfootR
 Type: Package
 Title: Dairy Farm Carbon Footprint Assessment
-Version: 0.1.2
+Version: 0.1.3
 Authors@R: 
     person("Juan", "Moreno", email = "juanmarcosmoreno@gmail.com", role = c("aut", "cre"))
 Description: Calculates the carbon footprint of dairy farms based on methodologies of the International Dairy Federation and the Intergovernmental Panel on Climate Change. Includes tools for single-farm and batch analysis, report generation, and visualization. Methods follow International Dairy Federation (2022) "The IDF global Carbon Footprint standard for the dairy sector" (Bulletin of the IDF n° 520/2022) <doi:10.56169/FKRK7166> and IPCC (2019) "2019 Refinement to the 2006 IPCC Guidelines for National Greenhouse Gas Inventories, Chapter 10: Emissions from Livestock and Manure Management" <https://www.ipcc-nggip.iges.or.jp/public/2019rf/pdf/4_Volume4/19R_V4_Ch10_Livestock.pdf> guidelines.
