+++ NEWS.md
@@ -1,12 +1,3 @@
 # cowfootR 0.1.3
 
-# cowfootR 0.1.2
-
-# cowfootR 0.1.1
-
-# cowfootR 0.1.0
-
-* Initial release
-* Dairy farm carbon footprint calculation functions
-* Batch processing capabilities
-* Comprehensive LCA methodology implementation
+* Added a `NEWS.md` file to track changes to the package.
