### 2026-01-13 f91072b
Docs: rename vignettes + update NEWS; fix inputs helpers
