### 2026-01-14 e296b90
Fix: explicit annual units (kg CO2eq yr-1) in returned emission objects
