### 2026-02-08 #724 @maurolepore
Hola @juanmarcosmoreno-arch,

Espero todo este bien, pero entendemos perfectamente que tus prioridades pudieron haber cambiado.

Como mencione arriba, si necesitas pausar el proceso podemos aplicar la etiqueta de *holding* por hasta un año. Sin embargo, aun en *holding* necesitamos tu confirmación ahora y cada tres meses.

Si por otra parte no recibimos respuesta el proceso que aplica no es *holding* sino que corresponde cerrar el issue.

> Si quien envió el paquete no ha solicitado una etiqueta de espera, pero no responde a las consultas, podremos cerrar el issue en el plazo de un mes desde el último intento de contacto. Este intento incluirá un comentario etiquetando a la persona, pero también un correo electrónico utilizando la dirección de correo electrónico que aparece en el archivo DESCRIPTION del paquete. Este es uno de los raros casos en los que quien hace la edición intentará ponerse en contacto con el autor por correo electrónico.
>
> -- https://devdevguide.netlify.app/es/softwarereview_policies.es
