### 2026-02-08 #724 @juanmarcosmoreno-arch
Estimado @maurolepore primero que nada pido disculpas por no haber dado una respuesta. Avance con muchas de las observaciones que hicieron y fueron muy útiles pero no pude completarlas aún. Solicito que se aplique la etiqueta holding para retomarlo más tarde. Desde ya muchas gracias y vuelvo a disculparme.
