### 2026-02-09 #724 @maurolepore
Menciono a @paocorrales, @NPJuncal, para que esten al tanto. En resumen, vamos a poner el proceso en suspenso y considerar retomarlo cada tres meses por hasta un máximo de 1 año.

Cuando los cambios esten listos les voy a pedir analizarlos para ver si son suficientes o necesitan mas cambios. Cualquier problema de disponibilidad lo consideraremos en su momento, y caso a caso.

Muchas gracias por su trabajo hasta ahora y por su entendimiento.
