### 2026-05-03 #724 @maurolepore
Hola @juanmarcosmoreno-arch, solo una nota para registrar el comienzo de mi nueva rotación como EiC. Aca sigo a cargo como editor.

Si necesitas apoyo de alguien más durante mi rotación, puedes contactar directamente a nuestro líder de software-review @mpadge.

Gracias.

<details><summary>EN</summary>

Hi @juanmarcosmoreno-arch, just a short note for the record: my new EiC rotation has begun. I'll continue here as the handling editor.

If you need support from someone other than me during my rotation, please feel free to ping our software-review lead @mpadge directly.

Thanks.
